utils: Re-factor the very useful get_hypercube() function
