+++ orangecontrib/spectroscopy/widgets/owstackalign.py
@@ -16,6 +16,8 @@ from Orange.widgets import gui, settings
 from orangecontrib.spectroscopy.widgets.owhyper import index_values, values_to_linspace
 from orangecontrib.spectroscopy.data import _spectra_from_image, getx, build_spec_table
 from orangecontrib.spectroscopy.widgets.gui import lineEditIntRange
+from orangecontrib.spectroscopy.utils import ⁠\
+    NanInsideHypercube, InvalidAxisException, ⁠get_hypercube
 
 
 # the following line imports the copied code so that
@@ -74,38 +76,8 @@ def alignstack(raw, shiftfn, ref_frame_num=0, filterfn=lambda x: x):
     return shifts, aligned
 
 
-class NanInsideHypercube(Exception):
-    pass
-
-
-class InvalidAxisException(Exception):
-    pass
-
-
 def process_stack(data, xat, yat, upsample_factor=100, use_sobel=False, ref_frame_num=0):
-    ndom = Domain([xat, yat])
-    datam = Table(ndom, data)
-    coorx = datam.X[:, 0]
-    coory = datam.X[:, 1]
-
-    lsx = values_to_linspace(coorx)
-    lsy = values_to_linspace(coory)
-    lsz = data.X.shape[1]
-
-    if lsx is None:
-        raise InvalidAxisException("x")
-    if lsy is None:
-        raise InvalidAxisException("y")
-
-    # set data
-    hypercube = np.ones((lsy[2], lsx[2], lsz)) * np.nan
-
-    xindex = index_values(coorx, lsx)
-    yindex = index_values(coory, lsy)
-    hypercube[yindex, xindex] = data.X
-
-    if np.any(np.isnan(hypercube)):
-        raise NanInsideHypercube(np.sum(np.isnan(hypercube)))
+    hypercube, lsx, lsy = get_hypercube(data, xat, yat)
 
     calculate_shift = RegisterTranslation(upsample_factor=upsample_factor)
     filterfn = sobel if use_sobel else lambda x: x
